+++ src/FractalModel.hs
@@ -6,17 +6,17 @@ import Viewer
 
 type FractalRef = [[FComplex]]
 
-screen :: Float -> Float -> (FComplex -> FComplex) -> FractalRef
+screen :: Double -> Double -> (FComplex -> FComplex) -> FractalRef
 screen w h f = map (\x -> [ f $ toCart w h (x :+ y) | y <- [0..h-1] ]) [0..w-1]
 
-toCart :: Float -> Float -> FComplex -> FComplex
+toCart :: Double -> Double -> FComplex -> FComplex
 toCart w h (x :+ y) = r*(2 * x / w - 1) :+ (1 - 2 * y / h)
   where r = w / h
 
-oZoom :: Float -> FComplex -> FComplex
+oZoom :: Double -> FComplex -> FComplex
 oZoom z (x :+ y) = (x/z) :+ (y/z)
 
-offsets :: Float -> Float -> FComplex -> FComplex
+offsets :: Double -> Double -> FComplex -> FComplex
 offsets rx ry (x :+ y) = (x+rx) :+ (y+ry)
 
 data FractalModel
